+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - GST for quantum instruments
 - mcm_results property for Circuit object
 - qcal -> pyquil transpiler
+- PostProcessor class for better handling of post-processing results with 
+optional passes
 
 ### Changed